maj sur le fork officiel linuxmao

+++ sources/Ports.cpp
@@ -104,49 +104,49 @@ rtosc::Ports rtCloud {
     {"numGrains:f", rDoc("NumGrains"), nullptr,
      [](const char *msg, rtosc::RtData &data) {
          SceneCloud *sceneCloud = (SceneCloud *)data.obj;
-         int arg = rtosc_argument(msg, 0).f;
+         float arg = rtosc_argument(msg, 0).f;
          sceneCloud->cloud->setNumGrains(arg);
      }},
     {"direction:f", rDoc("Direction"), nullptr,
      [](const char *msg, rtosc::RtData &data) {
          SceneCloud *sceneCloud = (SceneCloud *)data.obj;
-         int arg = rtosc_argument(msg, 0).f;
+         float arg = rtosc_argument(msg, 0).f;
          sceneCloud->cloud->setDirection(arg);
      }},
     {"windowType:f", rDoc("WindowType"), nullptr,
      [](const char *msg, rtosc::RtData &data) {
          SceneCloud *sceneCloud = (SceneCloud *)data.obj;
-         int arg = rtosc_argument(msg, 0).f;
+         float arg = rtosc_argument(msg, 0).f;
          sceneCloud->cloud->setWindowType(arg);
      }},
     {"spatialMode:f", rDoc("SpatialMode"), nullptr,
      [](const char *msg, rtosc::RtData &data) {
          SceneCloud *sceneCloud = (SceneCloud *)data.obj;
-         int arg = rtosc_argument(msg, 0).f;
+         float arg = rtosc_argument(msg, 0).f;
          sceneCloud->cloud->setSpatialMode(arg, -1);
      }},
     {"midiChannel:f", rDoc("MidiChannel"), nullptr,
      [](const char *msg, rtosc::RtData &data) {
          SceneCloud *sceneCloud = (SceneCloud *)data.obj;
-         int arg = rtosc_argument(msg, 0).f;
+         float arg = rtosc_argument(msg, 0).f;
          sceneCloud->cloud->setMidiChannel(arg);
      }},
     {"midiNote:f", rDoc("MidiNote"), nullptr,
      [](const char *msg, rtosc::RtData &data) {
          SceneCloud *sceneCloud = (SceneCloud *)data.obj;
-         int arg = rtosc_argument(msg, 0).f;
+         float arg = rtosc_argument(msg, 0).f;
          sceneCloud->cloud->setMidiNote(arg);
      }},
     {"active:f", rDoc("Active"), nullptr,
      [](const char *msg, rtosc::RtData &data) {
          SceneCloud *sceneCloud = (SceneCloud *)data.obj;
-         int arg = rtosc_argument(msg, 0).f;
+         float arg = rtosc_argument(msg, 0).f;
          sceneCloud->cloud->setActiveState(arg);
      }},
     {"locked:f", rDoc("Locked"), nullptr,
      [](const char *msg, rtosc::RtData &data) {
          SceneCloud *sceneCloud = (SceneCloud *)data.obj;
-         int arg = rtosc_argument(msg, 0).f;
+         float arg = rtosc_argument(msg, 0).f;
          sceneCloud->cloud->setLockedState(arg);
      }},
 };
@@ -155,13 +155,13 @@ rtosc::Ports nonRtCloud {
     {"x:f", rDoc("X"), nullptr,
      [](const char *msg, rtosc::RtData &data) {
          SceneCloud *sceneCloud = (SceneCloud *)data.obj;
-         int arg = rtosc_argument(msg, 0).f;
+         float arg = rtosc_argument(msg, 0).f;
          sceneCloud->view->setX(arg);
      }},
     {"y:f", rDoc("Y"), nullptr,
      [](const char *msg, rtosc::RtData &data) {
          SceneCloud *sceneCloud = (SceneCloud *)data.obj;
-         int arg = rtosc_argument(msg, 0).f;
+         float arg = rtosc_argument(msg, 0).f;
          sceneCloud->view->setY(arg);
      }},
     {"xExtent:f", rDoc("XExtent"), nullptr,
